Ensure that additive devices never have CMYK separations.
For some devn devices, we automatically add new colorants
as we meet them (e.g. psdcmyk).

For RGB ones, we should NEVER add the CMYK separations if
we encounter them. This can happen if we have a DeviceN
colorspace that uses an explicit /Cyan colorant.

This is mostly to match the behaviour of the PDF interpreter.
The PDF interpreter scans the resources used on a page to
count the number of spots that that page uses. It chooses
to ignore the standard CMYK colorants for the purpose of this
count, hence if the core of ghostscript does not do the same
we can get out of step.

It would be very hard to modify the PDF interpreter to
selectively ignore such colorants, as to do it properly
it would need to take account of both what device we are going
to (which would make the pdfinfo returned dictionary device
specific), and where those resources are used (as using them
in a DeviceCMYK blending space should NOT add them). Easier
all around to just leave well enough alone.