pdfwrite - improver preservation of clip text rendering modes
This arises from a customer file, but we already knew the code was less
than optimal, and there were comments describing why.

Because of the way the old (PostScript) PDF interpreter worked, when we
drew text with a text rendering mode which involved clipping (4, 5, 6
and 7) we would write the text and then separately write a path and
clip to that path. This results in larger, slower, output and the
path is written at the resolution of the device. If pdfwrite was run at,
say, 72 dpi, then small text would end up creating a path which was
unusable at higher resolution.

With the new PDF interpreter we can revisit that, and this commit does
so.

The pdfwrite device doesn't write text until it needs to, which allows
it to often write a more compact version, and it also executes a q
(gsave) and Q (grestore) round it. We can't do that with text involving
a clip, as the grestore woould restore away the clip.

This commit alters the PDF interpreter so that it sends a spec_op to the
device, if it is preserving text rendering modes, when we switch to a
mode which involves a clip. The itnerpreter remembers this has been done
and when we restore back to a state where the text rednering mode does
not involve a clip it again sends a spec_op to the device to inform it
that the clip should be removed.

Note that the spec specifically forbids switching from a clipping mode
to a non-clipping mode.

In the pdfwrite device we set a flag whenever we get a spec_op telling
us that we have switched to a clipping mode. At that point we flush any
accumulated text and do a gsave. From then on we do not gsave/grestore
around text operations. When we are informed by the interpreter that
we have restored back to the point where the text mode is non-clipping
we again flush any accumulated text, and issue a grestore to match the
gsave.

This results in a few files in the test suite producing smaller output
but no actual differences.

The original customer test file is markedly improved when run to the
pdfwrite device at 72 dpi (to work around a bug in Apple Preview) and
the output file is somewhat smaller.