Coverity ID #69096 Heap-buffer overflow in text_to_stream
I'd made an error with the use of the viewer stat stack which could lead
to us trying to use a state past the end of the stack, in extreme cases.

On reflection there's really no need for this to be tracked like a
grpahics state parameter, so I've moved it into the general pdfwrite
device structure, which simplifies the code and solves the problem.