Bug 707790: Fix SEGV in transparency handling.
We have a file with a top level CMYK group. Into this we tile
a (supposedly) transparent pattern with some spot colors.

Accordingly, we set up a page level transparency group. This
creates a pdf14_ctx, and creates a buf within that ctx.

When we start to play the pattern back, the pattern code
'stashes' the pdf14 state. Then during playback, it thinks
that the number of components has changed, so it frees the
existing ctx, and makes a new one.

At the end of the playback, it puts the old ctx (now freed)
back, and this later gets freed again, causing a SEGV.

I experimented with a fix whereby I always NULLed the ctx after
we stored the pdf14 state, so it wouldn't get freed within
the pattern playback. This caused crashes because we actually
use this ctx in order to push a transparency group in some
cases.

Reviewing the ctx, I can't obviously see *why* it needs to be
freed, and indeed, tests with the cluster don't show any ill
effects from not freeing it. Maybe the need for this has gone
away over the intervening decade?

So, the fix here is just not to free the ctx.