OSS-Fuzz 69118: Fix referencing into a bad devn_params struct.
There are times when we shouldn't dereference the devn_params
directly, but should retrieve it using the device call. I don't
claim to understand this, but that's what the comments there
say.

Accordingly, use this call. Without that, we can get a random
number back for the black component, and that makes us access
into hyperspace.