PDF interpreter - handle null objects
This came up when investigating the output of pdfwrite for a custoemr,
it isn't the original problem but was noticed in passing. If we get an
empty object such of the form:

x y obj
endobj

then we were failing to deal properly with it, we would fall of the end
of the endobj and throw an error. We should simply create and push a
NULL object. This commit handles that. Of course a NULL object might not
be valid for a given operation, but that's a different problem.