pdfwrite - fix Optional Content with PDF output < 1.4
No bug report, the file is from a customer, but any PDF file with
optinal content is sufficient to demonstrate the problem.

Commit 	de0e9e886a45ab30d71a6ff3944c768d99f79767 added support for
optional content in pdfwrite and the PDF interpreter, but I forgot to
check the output level of the PDF file in pdfwrite. This could lead to
us writing Optional Content to a PDF file which was less than PDF 1.5,
which is where optional content was introduced.

In this commit, if we get optional content check the version of PDF
being output, if it's less than PDF 1.5, additionally check if we are
producing PDF/A. Depending on the conditions, either drop the Optional
Content or abort PDF/A processing. In either case warn the user.