PDF interpreter & pdfwrite - more work on Optional Content
Bug #707799 "problems with preserving optional content"

The file has three optional content groups, each of which is associated
with a Form XObject. This didn't work properly when producing PDF/A
or any PDF < 1.5.

It hadn't really ocurred to me that more work needed to be done when the
pdfwrite device is producing a PDF which does not support Optional
Content. Some work along those lines has al;reayd been done but
realistically this needed to be done at least in part in the PDF
interpreter.

This commit adds a new device parameter to pdfwrite 'WantsOptionalContent'
which is false when going to ps2write, or the version of PDF does not
support Optional Content.

In that case we no longer send the Optional Content information from
the intperpreter. I've updated various places in the pdfwrite code to
also check, just in case we somehow get Optional Content sent when we
should not.

Finally two additional bug fixes, there was a missing 'return' from the
spec_op handling for optional content which meant it returned a
spurious error. Harmless but annoying. We were not emitting pending
optional content for Form XObjects. Ordinarily we don't preserve Form
XObjects from PDF, so I hadn't considered this case, but when we have
a form with a transparency Group we *do* preserve the form. In that
case we also need to write the Optional Content to the Form.