pdfwrite - don't encrypt strings when writing to an ObjStm
Ordinarily we encrypt strings when the output PDF file is encrypted,
this is as per the specification. However we *don't * want to encrypt
the strings when we are writing to a stream, because the stream itself
is encrypted.

If we were using an ObjStm then we weren't noticing the fact that we
were writing to a stream and encrypted the strings when we shouldn't.
This showed up with the DocInfo dictioanry contents, in particular the
Producer string.