PDF interpreter - reword misleading warning, remove pointless retry
Bug #707787 " Lost elements when extracting a page"

When we tried to dereference a reference to an object decalred in the
xref as free we emitted a warning and allegedly tried to treat the next
object number as an offset.

In fact we didn't do that at all. We just tried to use the offset from
the xref, which should always be 0 and guaranteed to fail.

Reworded the message and just return a NULL object instead.