PDF interpreter - process previous XRefStm from modified hybrid files
Bug #707784 " Lost elements when extracting a page"

The problem with this file is that it has been incrementally modified
numerous (at least 6) times. On one occasion it has been linearised and
on at least two occasions saved as a hybrid file; th trailer for these
saves has both an xref and an XRefStm entry.

In addition, the xref tables from the hybrid files declare a large
number of objects to be free, while the XRefStm contains references for
real in-use objects. This is legal the specification declares these as
'hidden objects' or something similar. However, the file also uses
these objects. In particular it uses some graphics states which are
declared as free in the xref table.

This causes us to attempt to repair the PDF file and the result of
that is incorrect processing. Obviously this is always a risk when
attempting to repair a broken file.

The only way to deal with this 'properly' is to process all the XRefStm
entries. We deal with multiple incremental saves by descending to the
earliest xref, processing that, then moving back to the next least recent
and processing that (which updates any altered references from the earlier
xref) and so on. We were previously only dealing with an XRefStm if it
was present in the most recent trailer.

This commit processes any XRefStm from the trailer *after* the xref has
been processed and before rewinding one level in the stack of trailers.
So we now process any xref, then process any XRefStm, then move up one
level and repeat until we have processed all the trailers.

This commit also removes any XRefStm from the trailer dictionary before
merging it with the primary trailer dictionary, we don't want to retain
the XRefStm in the primary trailer once the xref is processed, because
then we'd try to process it again.