pdfwrite - handle multiple references to a stream pdfmark
Bug #707787 "Lost elements when extracting a page"

This turned up with some marked content in the specimen file. If we
have a PDF file as input, and the PDF file contains marked content, then
the PDF itnerpreter generates pdfmark operations to preserve that
content in the output.

It is possible for multiple marked content specifications to reference
the same stream object in the original file. If that happened then when
we tried to create the stream a second time we would throw an error.

Instead, we now treat an attempt to write a stream which already exists
as simply an additional reference. We note that we have a duplicate
stream so that when it is closed we can avoid an error there too.