pdfwrite - treat the HPGL stick font as a normal font for colour purposes
Bug 707807 "HPGL/2 stick font can be filled with the wrong colour"

Reported by a customer, the file is a much reduced portion of a Quality
Logic test file (29_01.bin, but differs significantly from the one we
have).

This fills a rectangle with fill type 10, 15% and then draws some text,
which appears to end up using the stick font.

Since commit 31753bd24be80831a7d1ee2da8a4d4f88d37cb40 we have been
writing the colour for the stick font as a stroke colour, but it seems
that even though the type 3 CharProcs execute a stroke, in PDF it uses
the current fill colour to draw the stroke.

So the PDF interpreter used the existing fill colour, which was a
Pattern space from the fill.

This commit just adds the stick font to the list of font types that we
handle by emitting both a stroke and a fill colour.