PCL & pdfwrite - ignore RasterOPs and warn user
The PCL RasterOP mechanism can't be replicated in the PDF imaging model
in the majority of cases. Previously we've simply ignored the problem
and proceeded normally, this leads to incorrect output, without warning.

In this commit we add detection of high level devices and their support
(or lack of) for RasterOPs. If the device does not support RasterOPs
then for PXL and *most* PCL/HPGL-2/RT-Raster operations we flag up a
warning on stdout that there's a potential problem and ignore the
RasterOP.

For the case of 'show_background_char' we abort the entire operation,
because this often (not always, but very often) is used as a clipping
operation, and ends up writing a rectangle which obscures the character
being drawn. By aborting the entire operation we get many more files
producing correct, legible text. A couple of files which *ought* to
draw a white bacground to the character no longer do so. These files are
specific tests for transparency RasterOPs and already convert
incorrectly anyway, they just do so slightly differently now.

There was also a missing pcl_grestore in an error path in the function
show_char_background which I've added in passing. I noticed it because
initially I copied that block of code when aborting the operation, and
discovered it made some files render incorrectly.

NB the value '252' is a magic number :-( Apparently this is the default
state for RasterOPs, which appears to be opaque/opaque/opaque for the
source/destination/texture channels. Because that's the default we
dont issue a warning or abort if setting it.