pdfwrite - undo commit 	4dcfae36bb4dcbc4ef3b5e5afc98bcde0d6b9ddc
Commit 4dcfae36bb4dcbc4ef3b5e5afc98bcde0d6b9ddc added detection for an
SMask when producing PDF/A-2 with ColorConversionStrategy set to a
device space other than DeviceGray. This was because the version of
veraPDF at the time gave an error on the SMask data.

It appears the current version of veraPDF does not do so, and nor does
the Preflight in the current version of Acrobat. I still cannot find any
authoritative source for whether this is permitted or not. In my opinion
it should be because the SMask is used as an alpha channel, we do not
actually draw anything in DeviceGray.

In the absence of any specific documentation, and given that standard
conformance checkers no longer complain, I'v put the code back the way
it was before so that we will carry the SMask through in DeviceGray. The
resulting PDF is conformant according to Acrobat Professional of this
date (2024.001.20604) and veraPDF 1.24.3.