pdfwrite - fix PDF/A co-ordinate clamping
Bug #707811 " Unreadable document due to font with GS 10"

The problem was introduced with commit
949be69c7039afa848095acbd6d1cb772ed8510c which was itself a fix for
commit 3156b0aa8ac4e9ef440c169799658da249eb41c9

Both of those commits relate to limiting the co-ordinate space we use;
originally limited to +/-16384, to work with Acrobat 4, which we
removed, then reinstated for PDF/A only, but with a limit of 32768
because the PDF/A specification pickles the limitations of Acrobat 5
into the spec.

There was a small error in one of the checks; we pass in a rectangular
bounding box of a path to see if we need to scale it. The checking
arithmetic was using the bbox values verbatim, but in fact these are
fixed precision 24.8 values, not plain floating point.

This resulted in us scaling paths 1/256th the maximum size.

The example file has a type 3 font with a FontMatrix scaling the font
co-ordinate space by 1000, this was enough to trigger the rescaling
when it should not have, resulting in the output being incorrectly
scaled.