Don't include the NULL terminator in a param string
For some reason, the opvp/oprp device get_params method adds one to the
value of strlen() when return it's parameter strings, even though
gs_param_string contents are not null terminated. That causes the length
validation of the "Driver" parameter to fail, and to throw a spurious error.

So remove the "+ 1" for that param string. I've left the rest of the strings
as is until I have time to do more comprehensive testing.