pdfwrite - Disallow production of PDF/A-2 with device-independent colour
No bug report for this, I stumbled over it while working on soemthing
else.

PDF/A-2 requires that, if there is no OutputIntent, all objects at the
page level shall have a /Group with a /CS entry defining the blending
space.

We have no way to enforce that in pdfwrite, and no way to check if it
is the case or not. In fact, in general, I believe it will not be the
case. That basically means we can't create a PDF/A-2 file using the
-sColorConversionStrategy=UseDeviceIndependentColor control.

This commit checks to see if we are producing PDF/A, at a level which
permits transparency (A-2 or better) and the colour conversion strategy
is device-independent colour. If so it uses the PDFACompatibilityPolicy
to decide what to do. We cannot 'ignore' the colour conversion strategy
in this case, so we fall back to a normal PDF file. In all cases we emit
a warning.