pdfwrite - emit ToUnicode CMaps for simple fonts with standard glyphs
Bug # 707827 "FontType3 fnt defined in file input not mapped to unicode"

When a simple font (not a CIDFont) is entirely composed of glyphs with
names found in one of the standard Encodings, there is no need to write
a ToUnicode CMap, as the glyph names already contain the information.
So pdfwrite checked every glyph in a font's encoding and if all the
glyph names were standard it did not emit a ToUicode CMap.

However some PDF consumers seem to be unable to use the glyph names,
and only use the ToUnicode CMap.

So this commit removes the check on the encoding and, providing all the
other criteria are met, emits ToUnicode CMaps for these fonts.

This is a change in behaviour, and the resulting PDF files are slightly
larger than previously, so there is a new control ToUnicodeForStdEnc
which defaults to true and can be used to restore the old behaviour.