Bug 707830: Change 'num_std_colorants' when we change color space.
pdf14 devices can change their color model due to different groups.
This happens on "pdf14_push_color_model" and "pdf14_pop_color_model"
calls respectively.

At the moment, nothing directly records the fact that the color
model has been changed. This means that calculations based upon
the expected number of 'standard' colorants present can go wrong.

The example in the bug is of an rgb+spots device, that is almost
immediately 'pushed' into being a cmyk+spots device (to cope with
a CMYK page group).

We then use a pattern that is rendered using a pattern-clist. The
clist is written using num_components == 15 (calculated as 11
page_spot_colors + 4 process components). This is correct.

When the clist is played back, the compositor operations recalculate
the number of components based upon the 'num_std_components' the
device has - which was originally 3. This means num_components
ends up as 14, and the mismatch causes an error.

We implement a fix here in that we update 'dev->num_std_colorants'
on a push, and replace it on a pop. We tweak the calculations to
use this, rather than the devn_params ones.

This seems to be harmless with regards to testing in the cluster,
solves the bug, and seems to be sensible.