PDF interpreter - treat invalid CMaps as a warning
Bug #707831 " I found that after the scanned PDF was compressed, the text in the PDF Ctrl+A, Ctrl+C, and Ctrl+V were garbled in the text editor."

The problem is that the original file has multiple invalid ToUnicode
CMaps. Acrobat seems to ignore the fact that these are invalid, and the
preflight syntax analysis in Acrobat also fails to edtect the problem.

Because we do detect the error we don't read the ToUnicode CMaps and
so Unicode code points are then not available for pdfwrite to create
ToUnicode CMaps in the output PDF file.

The ToUnicode CMaps contain multiple entries like:

1450 beginbfchar
<00ff> <2014>

Adobe Technical Note 5014 states quite clearly (page 71):

"The beginbfchar and endbfchar operators map int number of individual
input codes (srcCode) to a corresponding number of individual character
codes (dstCode) or character names (dstCharname), where int can be = 100."

Given that the creator is the Founder RIP (which is important in China)
and Acrobat clearly ignores the error, this commit permits the number
of code maps to grow until the stack is exhausted, though it will give a
warning if the number exceeds the specified limit.

This exposed a few problems in the code to increase the size of the
stack for the basic CMap interpreter, which are also fixed here.