Coverity ID 427562 - prevent underflow
I *think* Coverity is correct here, there is a lack of error checking
in the decoding which assumes that if we get a command, or partial
offset there will always be at least 1 byte available in the 'row' data.

I think that's not a valid assumption in the case of corrupted data. I'm
not certain what the 'correct' approach should be, since this function
does not appear to be able to signal an error, so I've chosen to mimic
the code for 'partial_bytecount' which exits with end_of_row true.

I expect this will lead to garbage output, but then the data must be
corrupted to hit this state.