Coverity 427623: Add limit checks for CMap stacks
It's not all clear to me what Coverity is complaining about, so this may be
the first of a few attempts.

This just adds a hard limit check on the size of the stack before grabbing
the maps from it.

This is, essentially, redundant, as it's the same limit the stack extension
code uses, but I'm theorising that is too removed for coverity to understand.