PDF interpreter - Totally flush the pattern cache on page completion
Bug #707879 " image not rendered resp. in bad position by merging pdf"

The problem here is to do with the pattern cache, pdfwrite and the
PostScript interpreter graphics state.

At the end of each page we flush any patterns in the pattern cache, in
case we are running under the PostScript environment (Ghostscript as
opposed to GhostPDF). We don't want any of our PDF objects to end up
in the PostScript world.

But... When the pdfwrite device is active we don't add real patterns to
the pattern cache, but we do add dummy entries. These are *not* removed
by the call to gx_pattern_cache_winnow(). In fact there is no mechanism
for removing these at all, other than freeing the cache.

So when we start the second file we see the pattern it uses, and it
has the same ID as the pattern from the first file (because we reset
everything between files, so we restart the ID allocation). We then
discover the dummy entry in the pattern cache, and we reuse that for
the first Pattern. But of course, that is not the same pattern at all
so it fails to render correctly.

The obvious 'fix' of flushing the dummy entries doesn't work, that ends
up in (apparently) infinite looping with some PostScript input files. I
presume this is why we don't flush the dummy entries!

So I have added a new function to explicitly flush the entire cache,
including any dummy entries, and called that in place of the winnow()
function. This leaves the PostScript behaviour unchanged, but fixes the
problem specific to the pdfwrite device when the input is PDF.