OSS-fuzz 70522 - fix NULL dereference
This is fallout from the error handling change.

Previously we used the return value from pdfi_dict_get_type() to avoid
dereferencing a NULL Subtype, because that would return an error. But
we now make the code 0 if we don't stop because of the error.

That means we try to use pdfi_name_is() on a NULL object, and that
code dereferences the object unconditionally (the caller is supposed
to ensure the name object is valid).

Instead of using 'code >= 0' use 'Subtype != NULL' to guard against
this.