PDF interpreter - fix compiler warning
We used to use pdfi_pattern_purge_all_proc() to force the pattern cache
'winnow' function to free all the entries, but it turned out that the
winnow function doesn't free 'dummy' entries, and we need to free those
as well, so we moved to a new 'flush' function for the cache.

This left the pdfi_pattern_purge_all_proc function unused and gcc
complains, so remove it permanently.