mswinpr2 device - fix subclassing
Bug #707909 "FirstPage to LastPage command line"

The mswinpr2 device works in a weird way by calling the prn device
'open' method from its own 'open' device method. We have some code to
deal with the consequences of this with device subclassing (the
detection of devices copied using copydevice needs adjusting) but it
didn't cater for modifying the device contents.

The upshot is that we ended up modifying lpfnCancelProc and hDlgModeless
(when NOCANCEL is not selected) in the wrong device structure, the
subclassing device, not he real mswinpr2 device. This then meant
that these values were incorrect, and illegal, when printing a page
which caused an error.

Just make sure to use the correct device for this stuff after we've
reopened the device.