PDF interpreter - fix memory leak with XRef streams and repairing files
Specimen file is bug #695403 "shout072-xref disabled.pdf"

When we process an xref stream, we assign the context's Trailer member
to the dictionary from the stream. We need to countdown the Trailer
member before we do that, in case we are repairing a badly broken file
and have already assigned a dictionary to the Trailer.