PDF interpreter - fix a 'typo'
When finalizing the PostScript object containing the PDF interpreter
context; if the ps_stream isn't NULL (error occured) then we need to
'detach' it from the PDF context.

We were setting the PDF stream object to NULL instead of the member of
the structure which points to the PostScript stream. That meant we
leaked the PDF stream object.