PDF interpreter - fix error handling with document view preservation
Some errors were returning immediately, instead of jumping to the
'exit', which meant that various objects were not counted down.

Specimen file: Bugzilla_pdf\0688419_01832_w_a.PDF

The OpenAction is an indirect reference to object 61, but the file only
contains 60 objects, resulting in a 'undefined' error, and a return
without counting down the PDF object for the PageMode.