pdfwrite - fix negative array offset
OSS fuzz #70967

The initial problem is caused by a signed/unsigned mismatch. the
pdf_font_resource_t member 'count' is an unsigned int, and we were
passing 'chars_count' as a signed int. Because the signed int was
promoted to an unsigned in the comparison of the two, this passed.

We then tried to use the signed int in a memset leading to the address
sanitizer error.

Fix that by turning ghe signed integer into unsigned. The original
value is actually a uint64_t, but making it fully consistent would be
'awkward'.

This then led to an integer overflow, because we add 7 and divide by 8
to 'round up' the chars_count and pfres->count. If either were within
7 of the maximum uint then we would overflow which again would
potentially lead to an incorrect calculation in pdf_resize_array, which
is supposed to be guarded against by the comparison of the two values.

Add a specific test to ensure that neither can overflow.