PDF interpreter - fix memory leak with oddly broken files
bug #705725, pdf_loop_detect-62-pdfi_loop_detector_add_object.pdf

This one is hard to explain. A PDF file broken in certain ways could
lead us into a circular reference by referencing the circular objects
separately so that they both end up being dereferenced.

In this case the Info key in the Catalog points to an objct which is
in fact a Page dictionary. That dictionary has a Contents entry pointing
to object 3, and a Resources entry pointing to object 2.

When we 'sanitise' the Info dictionary we dereference the objects in
what we think is the right dictionary. So we end up dereferencing
the Contents. This (correctly) returns a circular reference error,
because object 3 is not a Resources dictionary at all, it's part of the
Pages tree, so it points to a dictionary which has a Kids entry. The
Kids entry is to object 2, which is part of a number tree, which has a
Kids array pointing to object 3. Circularity.

However, although we detect the circular reference, we still end up
storing a dereferenced object in the array.

Later we 'sanitise' the Resources entry, which goes directly to object 2
and does not detect a circular reference, because we already *have*
an object 2.

This leads us to a circular reference between objects 2 and 3.

To fix this, if we get an error while 'sanitising' array or dictionary
objects, replace the offending object with an indirect reference
and return the error.

This prevents the problem.

However, the file goes on to exhibit more leaks. This time there is an
error in the 'resolve_indirect' functionality which performs much the
same function as the 'sanitise' function, but for pdfmarks. In one
case we were storing the object in the array before recursively
checking it for circularity, which led to us storing a circularly
referencing object. Fix that by checking first, just as we do for
dictionaries already.

The second one was more subtle, if we found a circular reference then
we would leave the reference as an indirect object, but we would swallow
the error. For complicated reasons much like those above, this still
leads to circular references. Change the behaviour to return an error.

This makes more sense to me, we can't write a valid pdfmark if we have
an indirect object so we shouldn't try.