PDF interpreter - fix memory leak on filter to buffer error
specimen file OSS-fuzz #42309

When converting a PDF stream to an in-meory buffer we apply a
SubFileDecode filter to the 'base' stream so we can stop at the end of
the object.

But we apply that filter separately to the Filters for the stream which
means we also need to 'undo' it separately as well. We were failing to
do that if the routine to open the relevant filters returned an error.