PDF interpreter - use correct routine for freeing functions
File from OSS-fuzz #53758

In some places we simply weren't freeing the function when an error
occured, in others we were simply freeing the params.Function object.
This is sufficient for simple functions, but not when the function is
an array of functions. Instead use the graphics library routine to
free functions, which correctly calls hte right routine to clean up.