PDF interpreter - avoid assigning name if space did not change
file from OSS-fuzz #70529

The problem here is that we end up assigning a name to the initial
PostScript colour space. That space is not released until we close down
the PostScript interpreter, which is much too late to release the
name we assigned to it, and so it leaks.

This occurs because the very first colour space we try to use is an
invalid ICC space, with an invalid Alternate, but /N of 1. We fall right
back to creating a DeviceGray colour space from that.

If we actually created a new DeviceGray space this would still work out,
because we would release the colour space at some future time before
we returned to PostScript. However there's an optimisation in the
colour space assignment code which avoids assigning a new space if it's
the same as the existing space, which means we just leave the PostScript
DeviceGray in place.

I had thought that we might need to plumb a 'assign_name' boolean
through the entire colour space handling, but in fact all that is really
needed is to perform the same optimisation when assigning the name; if
the space hasn't changed we don't need to assign a name to it.