PDF interpreter - improve handling of integer overflow
Bug #707943 " Pdfwrite generates a blank page in a specific case since v10.02.0"

The problem is that the PDF file has a Form XObject with a /BBox where
the values are out of range for 32-bit integers. The implementation
limits in the PDF specification suggest that integers are limited to
32-bits.

Although we store values internally as signed 64-bits, we still flag a
warning for out of spec values, and clamp the values to a signed integer
for negative values. But we allow unsigned integers to reach the
maximum unsigned int before overflowing.

Two values in the BBox are out of range for signed integers, but the
other two, while out of signed integer range, are in range for unsigned
integers.

This caused a degree of confusion because we were creating an
intermediate value as a signed integer.

Fixed by changing the intermediate variable to an unsigned int, which
then requires some casting.