PDF interpreter - Dereferencing an object, multiple returns is an error
File from Bug #70577

The file ends abruptly, with a partial dictionary, ending with (more or
less) an indirect reference. We attempt to repair the file and rebuild
the xref, then carry on.

However, when we come to try and read the Resources dictionary (which is
the truncated object) we end up leaving all the dictionary contents on
the stack, the final object being the trsailing indirect reference.

This leads to us returning an indirect reference when we've been told
to dereference an indirect reference, and we cache the object as well.
Returning an indirect reference causes a memory leak because we
(eventually) end up creating a circular reference, involving the
indirect object returned from the cache.

When we dereference an object we only ever expect to see a single
object. The only way to end up with multiple objects is if the object
is an array, dictionary, or stream, and the object was truncated in
some fashion (spurious endobj or EOF). We could (as I note in the
comment) try to repair the object by walking back along the stack and
'closing' any open arrays or dictionaries. However it seems unlikely
that this would fix many files. If it seems useful we can do that in
future.

In the meantime, treating it as an error, and flushing the stack,
resolves the memory leak that otherwise occurs.

I also noted a mistake with a 'return code' instead of jumping to the
error label and properly cleaning up. It doesn't affect this file, and
I don't have an example file to trigger it, but let's fix it before it
does become a problem.