PDF interpreter & pdfwrite - don't write Metadata for Marked Content
Bug #70788 " Regression: Conversion to PDF/A-3b fails VeraPDF Validation, clause 6.6.2.3.1"

This is not, exactly, a regression. Older versions of the pdfwrite
device and the PDF interpreter did not preserve Marked Content, new
versions do. The problem arises because the Marked Content Properties
dictionary associated with a /BMC contains a /Metadata key. In order
for that to be valid we would haev to extend the XMP metadata and we
don't.

My first thought was to use Acrobat to convert the file to PDF/A-3 and
then see how it modified the XMP metadata. Amusingly, it doesn't and the
file it produces fails VeraPDF validation in exactly the same way. This
does make me wonder if the VeraPDF test is actually correct.

However, in order to work around the problem, I've modified the pdfwrite
device so that if we are producing PDF/A and we get a /BMC marked
content pdfmark, and the Properties dictionary contains a /Metadata key,
then we do the usual PDFACompatibilityPolicy dance to decide what to
do.

However, because of the way that these constructs are assembled using
pdfmarks, this still ends up leaving the actual Metadata written to the
file. It isn't referenced, and as a result VeraPDF is happy to have it
there, but it is wasteful. Nothing we can do about that if the source
is PostScript, but when the source is PDF we can do better, so....

Modify the PDF interpreter so that it knows about PDFA, PDFX and
PDFACompatibilityPolicy. When we encounter a BMC in the input PDF file,
examine its Properties dictionary to see if it contains a Metadata key.
If it does, delete the key/value pair from the dictionary and continue.
This means we don't write the Metadata itself to the file.