PDF interpreter - Type 4 functions, allow ridiculous decimal places
No file, the original is from a customer. The type 4 calculator function
is:

{dup 0 mul exch dup 0.7299999999999999 mul exch dup 0.8899999999999999 mul exch 0 mul}

Obviously 16 places of decimal accuracy is silly. It also would
overflow an internal buffer so we throw an error.

We only allow 16 bytes of numeric data, which is plenty for the maximum
integer (2^32), and even for fractional values with a large mantissa
allows for 5 decimal places of accuracy. As the mantissa grows larger
it's reasonable to assume the fractional part is less significant and
requires less accuracy.

There is (obviously) no limit to the level of accuracy of expression of
decimal numbers, but we need to limit it as we have a fixed internal
buffer. We could increase the buffer size but there's no value which
is guaranteed to be enough.

Instead, if we exceed the buffer size, and the number is a floating
point value, just drop any digits after we've filled the internal
buffer.

I doubt any loss of accuracy will be detectable.

I noticed in passing that the buffer size isn't using the #define'd
constant, which it should be, so I fixed that too.