pdfi: Memory leak with Pages circular reference
In the even that the /Pages dict /Count value doesn't match the number of
entries in the /Kids array, we try to derive a valid page count by
checking valid entries in the /Kids array.

In this case, the only entry in /Kids array is an indirect reference to the
/Pages dictionary, thus a circular reference.

There were two problems in this scenario:

Firstly, we allowed the array accessing code to replace the indirect object
array entry with the dereferenced object, thus circular reference is fully
realised.

Secondly, having spotted the circular reference, and throwing an error, we
weren't decrementing the reference counts on all the objects to which we were
holding references.

Fixed by not allowing the array accessor to replace and cache the object, but
doing so explicitly once we've validated the entry.

Secondly, decrementing the reference counts for the relevant objects before
throwing an error.